backend-v2/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
cloudinary
python-dotenv
python-slugify
redis
psycopg2-binary
slowapi
